add test for graph with views
